Moves auth into seperate module
