+++ meetings/auth/urls.py
@@ -1,0 +1,6 @@
+from django.conf.urls import url, include
+from auth import views
+
+urlpatterns = [
+    url(r'^', views.OsfAuthorizationCode.as_view()),
+]
